countries/requirements.txt
requests==2.31.0
boto3==1.33.9
amazon-dax-client==2.0.3